# Back/docs/swagger.yaml
basePath: /
definitions:
  DTO.ContentInfo:
    properties:
      body:
        type: string
      content_type:
        type: string
      created_by:
        type: integer
      createdAt:
        type: string
      end_date:
        type: string
      external_url:
        type: string
      extra_img_list:
        items:
          type: string
        type: array
      id:
        type: integer
      is_published:
        type: boolean
      main_img_url:
        type: string
      source:
        type: string
      start_date:
        type: string
      summary:
        type: string
      title:
        type: string
      updatedAt:
        type: string
    type: object
  DTO.ContentList:
    properties:
      content:
        items:
          $ref: '#/definitions/DTO.SummaryContent'
        type: array
      page:
        type: integer
      size:
        type: integer
      total:
        type: integer
    type: object
  DTO.CreateContent:
    properties:
      body:
        maxLength: 2000
        type: string
      device_id:
        items:
          type: integer
        type: array
      end_date:
        type: string
      img_list:
        items:
          type: string
        maxItems: 100
        type: array
      main_img:
        type: string
      start_date:
        type: string
      summary:
        type: string
      title:
        maxLength: 50
        minLength: 5
        type: string
    required:
    - device_id
    - main_img
    - title
    type: object
...
  DTO.ErrorResponse:
    properties:
      error:
        example: خطایی رخ داده است
        type: string
    type: object
...
  DTO.MessageResponse:
    properties:
      message:
        example: عملیات با موفقیت انجام شد
        type: string
    type: object
...
  DTO.SummaryContent:
    properties:
      created_at:
        type: string
      id:
        type: integer
      main_img:
        type: string
      summary:
        type: string
      title:
        type: string
    type: object
...
  DTO.UpdateContentRequest:
    properties:
      body:
        maxLength: 2000
        type: string
      end_date:
        type: string
      img_list:
        items:
          type: string
        maxItems: 100
        type: array
      main_img:
        type: string
      start_date:
        type: string
      summary:
        type: string
      title:
        maxLength: 50
        minLength: 5
        type: string
    type: object
...
  DTO.VitrinContentList:
    properties:
      items:
        items:
          $ref: '#/definitions/DTO.VitrineSlot'
        type: array
    type: object
...
  DTO.VitrineSlot:
    properties:
      content:
        $ref: '#/definitions/DTO.SummaryContent'
      is_auto:
        description: true یعنی این اسلات خودکار از آخرین اخبار پر شده
        type: boolean
      position:
        type: integer
      source:
        description: daralouWeb | daralouOperator
        type: string
    type: object
...
info:
  contact: {}
  description: مستندات API پروژه Back (احراز هویت، مدیریت محتوا، مدیریت دستگاه‌ها).
  title: Back API
  version: "1.0"
paths:
...
  /contents:
    post:
      consumes:
      - application/json
      description: ایجاد یک محتوای دستی (announcement) توسط اپراتور/ادمین
      parameters:
      - description: اطلاعات محتوا
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/DTO.CreateContent'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/DTO.MessageResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/DTO.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/DTO.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/DTO.ErrorResponse'
      security:
      - BearerAuth: []
      summary: ایجاد محتوای دستی
      tags:
      - Content
  /contents/{content_id}:
    get:
      description: دریافت جزئیات کامل یک محتوا؛ برای محتوای دستی، دسترسی دستگاه چک
        می‌شود (ادمین/اپراتور باید device_id را در query بدهد)
      parameters:
      - description: شناسه محتوا
        in: path
        name: content_id
        required: true
        type: integer
      - description: شناسه دستگاه (برای ادمین/اپراتور الزامی است)
        in: query
        name: device_id
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/DTO.ContentInfo'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/DTO.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/DTO.ErrorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/DTO.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/DTO.ErrorResponse'
      security:
      - BearerAuth: []
      summary: جزئیات یک محتوا
      tags:
      - Content
  /contents/{id}:
    delete:
      description: حذف یک محتوای دستی (محتوای اسکرپ‌شده قابل حذف نیست)
      parameters:
      - description: شناسه محتوا
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/DTO.MessageResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/DTO.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/DTO.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/DTO.ErrorResponse'
      security:
      - BearerAuth: []
      summary: حذف محتوا
      tags:
      - Content
    put:
      consumes:
      - application/json
      description: ویرایش یک محتوای دستی (محتوای اسکرپ‌شده قابل ویرایش نیست)
      parameters:
      - description: شناسه محتوا
        in: path
        name: id
        required: true
        type: integer
      - description: فیلدهای قابل تغییر
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/DTO.UpdateContentRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/DTO.MessageResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/DTO.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/DTO.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/DTO.ErrorResponse'
      security:
      - BearerAuth: []
      summary: به‌روزرسانی محتوا
      tags:
      - Content
  /contents/local:
    get:
      description: لیست صفحه‌بندی‌شده محتوای اختصاص‌داده‌شده به یک دستگاه؛ دستگاه
        (نقش device) خودکار محتوای خودش را می‌بیند، ادمین/اپراتور باید device_id را
        در query بدهد
      parameters:
      - description: شناسه دستگاه (برای ادمین/اپراتور الزامی است)
        in: query
        name: device_id
        type: integer
      - description: شماره صفحه (پیش‌فرض 1)
        in: query
        name: page
        type: integer
      - description: تعداد در هر صفحه (پیش‌فرض 20)
        in: query
        name: size
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/DTO.ContentList'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/DTO.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/DTO.ErrorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/DTO.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/DTO.ErrorResponse'
      security:
      - BearerAuth: []
      summary: لیست محتوای دستگاه
      tags:
      - Content
  /contents/scrap:
    get:
      description: لیست صفحه‌بندی‌شده اخبار اسکرپ‌شده (فقط اپراتور/ادمین)
      parameters:
      - description: شماره صفحه (پیش‌فرض 1)
        in: query
        name: page
        type: integer
      - description: تعداد در هر صفحه (پیش‌فرض 20)
        in: query
        name: size
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/DTO.ContentList'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/DTO.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/DTO.ErrorResponse'
      security:
      - BearerAuth: []
      summary: لیست اخبار اسکرپ‌شده
      tags:
      - Content
  /contents/vitrin:
    get:
      description: 'ویترین این دستگاه: آیتم‌های پین‌شده توسط ادمین در جایگاه خودشون
        + پرشدن خودکار اسلات‌های خالی با آخرین اخبار'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/DTO.VitrinContentList'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/DTO.ErrorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/DTO.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/DTO.ErrorResponse'
      security:
      - BearerAuth: []
      summary: لیست ویترین
      tags:
      - Content
...
securityDefinitions:
  BearerAuth:
    description: توکن دسترسی رو به‌صورت "Bearer {access_token}" وارد کن.
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"